+++ app/modules/wallet/models/entity/wallet_entity.go
@@ -9,7 +9,7 @@ import (
 type Merchant struct {
 	ID        string    `json:"id"`
 	UserID    string    `json:"user_id"`
-	Balance   float64   `json:"balance"`
+	Balance   int64     `json:"balance"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
 }
@@ -17,7 +17,7 @@ type Merchant struct {
 type Topup struct {
 	ID         string                      `json:"id"`
 	MerchantID string                      `json:"merchant_id"`
-	Amount     float64                     `json:"amount"`
+	Amount     int64                       `json:"amount"`
 	Status     paymentEntity.PaymentStatus `json:"status"`
 	CreatedAt  time.Time                   `json:"created_at"`
 	UpdatedAt  time.Time                   `json:"updated_at"`
